refactor(audit): split monolithic audit script into per-check modules
