+++ package.json
@@ -80,6 +80,10 @@
     "config/vite/vite.config.js",
     "scripts/a11y.js",
     "scripts/audit.js",
+    "scripts/audit/**/*.js",
+    "!scripts/audit/**/*.test.js",
+    "!scripts/audit/**/__snapshots__/**",
+    "!scripts/audit/test-utils.js",
     "scripts/audit-twig-stories.js",
     "scripts/check-node-version.js",
     "scripts/lib/cli.js",
